Handle ICP4D URL edge cases and refactor webview code
